fix limit of Int

+++ app/src/main/java/com/example/ilafedoseev/calculatorkotlin/Calculate.kt
@@ -45,7 +45,7 @@ class CalculateService {
     }
 
     /** Round the number */
-    fun roundNumber (num : Double) : Double {
+    fun roundNumber (num: Double) : Double {
         return Math.round (num * 10000.0) / 10000.0
     }
 
@@ -73,28 +73,28 @@ class CalculateService {
         }
     }
 
-    fun forEachToken(tokenArrayList : ArrayList<String>) : Double {
+    fun forEachToken(tokenArrayList : ArrayList<String>) : Float {
 
         var currentStr : String = ""
-        var currentNumber : Double = 0.0
+        var currentNumber : Float = 0.0f
 
         tokenArrayList.forEach {
             when {
                 it == "+" || it == "/" || it == "*" || it == "-" || it == "^" -> currentStr = it
 
                 currentStr == "-" -> {
-                    if (it != "") {currentNumber -= it.toDouble()}
+                    if (it != "") {currentNumber -= it.toFloat()}
                 }
 
                 currentStr == "/" -> {
                     if (it == "") {
                         currentNumber += 0
                     } else if (it == "0") {
-                        currentNumber = 0.0
+                        currentNumber = 0.0f
                     } else if (it.toDouble() > currentNumber) {
-                        currentNumber = currentNumber.div(it.toDouble())
+                        currentNumber = currentNumber.div(it.toFloat())
                     } else {
-                        currentNumber /= it.toDouble()
+                        currentNumber /= it.toFloat()
                     }
                 }
 
@@ -107,13 +107,13 @@ class CalculateService {
 
                 currentStr == "*" ->{
                     if (it != "") {
-                        currentNumber *= it.toDouble()
+                        currentNumber *= it.toFloat()
                     }
                 }
 
                 else -> {
                     if (!it.equals("")) {
-                        currentNumber += it.toDouble()
+                        currentNumber += it.toFloat()
                     }
                 }
             }
@@ -122,21 +122,21 @@ class CalculateService {
         return currentNumber
     }
 
-    fun getResult(number : Double): String {
+    fun getResult(number: Float): String {
         if (number > 0 && number < 1) {
-            return roundNumber(number).toString()
+            return roundNumber(number.toDouble()).toString()
         } else if (haveDot) {
             if (number % 1.0 == 0.0) {
-                return number.toInt().toString()
+                return number.toString()
             } else {
                 if (number.toString().length > 8) {
-                    return roundNumber(number).toString()
+                    return roundNumber(number.toDouble()).toString()
                 } else {
                     return number.toString()
                 }
             }
         } else {
-            return number.toInt().toString()
+            return number.toString()
         }
     }
 }
